Separate the schema from the individual script(s)
